fix(core): V3 durability + comprehensive test/doctest cleanup
V3 backend durability fixes:
- header.total_pages is now synced from the allocator before every flush;
  previously total_pages was never persisted and PageAllocator started
  from page 2 on reopen, overwriting existing node/edge pages.
- flush_to_disk() now writes a KV checkpoint (.v3checkpoint), records a
  WAL checkpoint, and truncates the WAL so KV data survives reopen.
- open() now calls WALRecovery::recover_kv() to restore KV data from
  WAL or checkpoint; KV store was never recovered previously.
- insert_edge validates node existence for both endpoints, returning
  SqliteGraphError::invalid_input("edge endpoints must reference
  existing entities") and matching the SQLite backend's behavior.

Test rewrites:
- edge_insertion_corruption_test.rs and node_slot_transaction_persistence
  rewritten against the GraphBackend public API; V3 uses B+Tree page
  storage, not the fixed 4096-byte V2 slot layout.
- Removed transaction_begin_corruption_proof.rs and
  json_parsing_diagnostics.rs (tested V2-only constructs/JSONCTX marker
  that no longer exist).
- graph_node_existence_enforcement.rs accepts the new V3 error string.
- phase32_cluster_pipeline_reconstruction_tests_clean.rs fixes a swapped
  name/file_path call and ensures an explicit flush before reopen.
- perf_gate_v32_tests.rs gracefully skips when Criterion data is absent.
- v3_algorithm_tests.rs updates wildcard patterns ("test" -> "test*")
  now that query_nodes_by_name_pattern implements wildcard matching.
- v3_name_index_tests.rs updates expectations for special-pattern
  substring matching.

Doctest cleanup:
- 59 pre-existing doctest failures fixed across algo/* and hnsw/*
  modules. Examples now compile as Result-returning blocks with valid
  identifiers (variables, public APIs, builders, config signatures).
  cargo test --doc reports 93 passed, 0 failed.

Public docs:
- sqlitegraph-core/README.md gains Safety Invariants, DSL Constraints,
  Performance & Instrumentation, and Schema Compatibility Matrix
  sections required by doc_tests.rs.
- clippy.toml + Cargo.toml bump MSRV to 1.89 (AVX-512 intrinsics).

+++ sqlitegraph-core/src/backend/native/v3/file_coordinator.rs
@@ -269,12 +269,12 @@ mod tests {
         // Write page 1
         let data1 = vec![1u8; 4096];
         coordinator.write_page(1, &data1).unwrap();
-        assert_eq!(coordinator.file_size(), V3_HEADER_SIZE as u64 + 4096);
+        assert_eq!(coordinator.file_size(), V3_HEADER_SIZE + 4096);
 
         // Write page 2
         let data2 = vec![2u8; 4096];
         coordinator.write_page(2, &data2).unwrap();
-        assert_eq!(coordinator.file_size(), V3_HEADER_SIZE as u64 + 8192);
+        assert_eq!(coordinator.file_size(), V3_HEADER_SIZE + 8192);
 
         // Read back
         let mut buffer = vec![0u8; 4096];
@@ -296,7 +296,7 @@ mod tests {
         let data = vec![42u8; 4096];
         coordinator.write_page(100, &data).unwrap();
 
-        let expected_size = V3_HEADER_SIZE as u64 + (99 * DEFAULT_PAGE_SIZE) + 4096;
+        let expected_size = V3_HEADER_SIZE + (99 * DEFAULT_PAGE_SIZE) + 4096;
         assert_eq!(coordinator.file_size(), expected_size);
 
         // Read it back
